+++ .gitignore
@@ -1,2 +1,4 @@
 .DS_Store
 node_modules
+npm-debug.log
+
